# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GachaPull - Pokemon & One Piece Gacha API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /packs:
    get:
      operationId: listPacks
      tags: [packs]
      summary: List all packs
      parameters:
        - in: query
          name: franchise
          schema:
            type: string
            enum: [pokemon, onepiece, all]
        - in: query
          name: isActive
          schema:
            type: boolean
      responses:
        "200":
          description: Packs list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pack"
    post:
      operationId: createPack
      tags: [packs]
      summary: Create a pack (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePackBody"
      responses:
        "201":
          description: Created pack
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pack"
...
  /packs/{packId}:
    get:
      operationId: getPack
      tags: [packs]
      summary: Get pack detail with card pool
      parameters:
        - in: path
          name: packId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pack with card pool
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackDetail"
    put:
      operationId: updatePack
      tags: [packs]
      summary: Update pack (admin)
      parameters:
        - in: path
          name: packId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePackBody"
      responses:
        "200":
          description: Updated pack
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pack"
    delete:
      operationId: deletePack
      tags: [packs]
      summary: Delete pack (admin)
      parameters:
        - in: path
          name: packId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /packs/{packId}/cards:
    post:
      operationId: addCardToPack
      tags: [packs]
      summary: Add card to pack pool (admin)
      parameters:
        - in: path
          name: packId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCardToPackBody"
      responses:
        "201":
          description: Card added to pack
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
    delete:
      operationId: removeCardFromPack
      tags: [packs]
      summary: Remove card from pack pool (admin)
      parameters:
        - in: path
          name: packId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cardId]
              properties:
                cardId:
                  type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # GACHA
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # Auth schemas
...
    Card:
      type: object
      required: [id, name, franchise, rarity, imageUrl, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        franchise:
          type: string
          enum: [pokemon, onepiece]
        rarity:
          type: string
          enum: [common, rare, super_rare, ultra_rare, legendary]
        imageUrl:
          type: string
        description:
          type: string
          nullable: true
        pullCount:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    Pack:
      type: object
      required: [id, name, franchise, priceCoins, imageUrl, isActive, cardCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        franchise:
          type: string
          enum: [pokemon, onepiece]
        priceCoins:
          type: integer
        priceUsd:
          type: number
          nullable: true
        imageUrl:
          type: string
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
        cardCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    PackDetail:
      allOf:
        - $ref: "#/components/schemas/Pack"
        - type: object
          properties:
            cardPool:
              type: array
              items:
                $ref: "#/components/schemas/PackCardEntry"
    PackCardEntry:
      type: object
      required: [card, probability]
      properties:
        card:
          $ref: "#/components/schemas/Card"
        probability:
          type: number
          description: Probability as decimal (0-1)
    CreatePackBody:
      type: object
      required: [name, franchise, priceCoins, imageUrl]
      properties:
        name:
          type: string
        franchise:
          type: string
          enum: [pokemon, onepiece]
        priceCoins:
          type: integer
        priceUsd:
          type: number
          nullable: true
        imageUrl:
          type: string
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
          default: true
    AddCardToPackBody:
      type: object
      required: [cardId, probability]
      properties:
        cardId:
          type: integer
        probability:
          type: number
          description: Probability as decimal (e.g. 0.05 = 5%)

    # Gacha schemas
...